pyproject.toml
[tool.poetry]
name = "user-bot"
version = "0.1.0"
description = ""
authors = ["nikita"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
django-environ = "^0.11.2"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
